add the clear function
